--- library/crypto-rand/src/nativeInterop/cinterop/sys_random.def
@@ -1,3 +1,0 @@
-package = org.kotlincrypto.random.internal
-headers = sys_random.h
-headerFilter = sys_random.h
